\name{AIC.Bsiland}
\alias{AIC.Bsiland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Akaike's An Information Criterion for siland package
}
\description{
The function gives the Akaike's 'An Information Criterion' for an object of class Bsiland.
}

\usage{
\method{AIC}{Bsiland}(object,...,k=2)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ an object of class Bsiland}
  \item{\dots}{Dots are not take into account in this version}
  \item{k}{not used}
}

\value{ AIC value
}
\references{
Sakamoto, Y., Ishiguro, M., and Kitagawa G. (1986). Akaike Information Criterion Statistics. D. Reidel Publishing Company.

}
\author{
O. Martin
}



