\name{dataCmoth}
\alias{dataCmoth}
\docType{data}
\title{
Codling moth counts
}
\description{
This data set contains measurements of codling moth counts in 54 orchards. Anonymized data issued from Ricci et al. (2011) study.
}

\usage{data("dataCmoth")}
\format{
  A data frame with 54 observations on 4 variables.


\tabular{rlll}{
    [,1] \tab X   \tab numeric \tab locations (anonymized) \cr
    [,2] \tab Y \tab numeric \tab locations (anonymized)\cr
    [,3] \tab trait    \tab numeric \tab number of treatments against codling moths in orchards\cr
    [,4] \tab Cmoth    \tab numeric \tab average count of collected larvae in cardboard traps wrapped around tree trunks.
  }
}


\references{
Ricci B., Franck P., Bouvier J.-C., Casado D. and Lavigne C. (2011). Effects of hedgerow characteristics on intra-orchard distribution of larval codling moth. Agriculture, Ecosystems & Environment. 140. 395-400. 10.1016/j.agee.2011.01.001. 
}



\keyword{datasets}
