\name{fitted.Fsiland}
\alias{fitted.Fsiland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extract Bsiland model Fitted values
}
\description{
Function extracting fitted values of an object of class Fsiland.
}

\usage{
\method{fitted}{Fsiland}(object,...)
}




%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{an object of class Fsiland}
  \item{\dots}{ Dots are not take into accounbt in this version}
}

\value{
Fitted values extracted from the object \command{object}.
}

\author{
O. Martin
}
