\name{fitted.siland}
\alias{fitted.siland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extract siland model Fitted values
}
\description{
Function extracting fitted values of an object of class siland.
}

\usage{
\method{fitted}{siland}(object,...)
}




%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{an object of class siland}
  \item{\dots}{ Dots are not take into accounbt in this version}
}

\value{
Fitted values extracted from the object \command{object}.
}

\author{
O. Martin
}
