\name{plot.Bsiland}
\alias{plot.Bsiland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot an object of class Bsiland
}
\description{
This function calls the function plot.lm() and plot the object result stored in x
}
\usage{
\method{plot}{Bsiland}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
an object of class Bsiland
}
  \item{\dots}{
the same options as for the function plot.lm()
}
}


