\name{plot.Fsiland}
\alias{plot.Fsiland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot an object of class Fsiland
}
\description{
This function calls the function plot.lm() and plot the object result stored in x

}
\usage{
\method{plot}{Fsiland}(x, ...)
}

\arguments{
  \item{x}{
  an object of class Fsiland

}
  \item{\dots}{
  the same options as for the function plot.lm()

}
}








