\name{plot.siland}
\alias{plot.siland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot an object of class siland
}
\description{
This function calls the function plot.lm() and plot the object result stored in x

}
\usage{
\method{plot}{siland}(x, ...)
}

\arguments{
  \item{x}{
  an object of class siland

}
  \item{\dots}{
  the same options as for the function plot.lm()

}
}








