\name{print.Bsiland}
\alias{print.Bsiland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Bsiland print
}
\description{
function printing an object of class Bsiland. It returns the model used and the estimated parameters.
}

\usage{
\method{print}{Bsiland}(x,...)
}


%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{an object of class Bsiland.}
  \item{...}{not take into account in this version}

}



%% ~Make other sections like Warning with \section{Warning }{....} ~


