\name{print.Fsiland}
\alias{print.Fsiland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Fsiland print
}
\description{
function printing an object of class Bsiland. It returns the model used and the estimated parameters.
}

\usage{
\method{print}{Fsiland}(x,...)
}


%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{an object of class Fsiland}
  \item{...}{not take into account in this version}


}


%% ~Make other sections like Warning with \section{Warning }{....} ~


