\name{print.siland}
\alias{print.siland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
siland print
}
\description{
function printing an object of class siland. It returns the model used and the estimated parameters.
}

\usage{
\method{print}{siland}(x,...)
}


%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{an object of class siland}
  \item{...}{not take into account in this version}


}


%% ~Make other sections like Warning with \section{Warning }{....} ~


