\name{residuals.Fsiland}
\alias{residuals.Fsiland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extract Fsiland Residuals 
}
\description{
Function extracting residuals for an estimated Bsiland model.
}

\usage{
\method{residuals}{Fsiland}(object,...)
}



%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{an object of class Fsiland}
\item{\dots}{Dots are not take into account in this version}
}
