\name{residuals.siland}
\alias{residuals.siland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extract siland Residuals
}
\description{
Function extracting residuals for an estimated siland model.
}

\usage{
\method{residuals}{siland}(object,...)
}



%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{an object of class siland}
\item{\dots}{Dots are not take into account in this version}
}
