\name{siland-package}
\alias{siland-package}
\docType{package}
\title{
\packageTitle{siland}
}

\description{
\packageDescription{siland} This package allows for analyzing the effect of landscape features on georeferenced point observations (described in a Geographic Information System shapefile format). It simultaneously estimates the spatial scales and intensities of landscape variable effects. It does not require any information about the scale of effect. It allows for effects tests, effects maps and models comparison.
\\

}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{siland}
\packageIndices{siland}

}

\author{
\packageAuthor{siland}
Maintainer: \packageMaintainer{siland}
}




\keyword{ package }

