\name{BIC.Bsiland}
\alias{BIC.Bsiland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Bayesian Information criterion
}
\description{
Compute the Bayesian Information Criterion for an object of class Bsiland
}

\usage{
\method{BIC}{Bsiland}(object,...)
}


%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ an object of class Fsiland}
\item{\dots}{Dots are not take into account in this version}
}

\value{
The BIC value
}

\author{
O. Martin
}
