\name{dataCmoth}
\alias{dataCmoth}
\docType{data}
\title{
Data set for codling moth counts
}
\description{
The dataframe dataCmoth gives measurements for codling moth 
counts in 54 orchards. X and Y give the locations for observed data. The number Cmoth is the mean count of collected
 larvae in cardboard traps wrapped around tree trunks. trait is a local variable that gives the number of traitements for codling moth in orchards.

}
\usage{data("dataCmoth")}
\format{
  A data frame with 54 observations on the following 4 variables.
  \describe{
    \item{\code{X}}{a numeric vector}
    \item{\code{Y}}{a numeric vector}
    \item{\code{trait}}{a numeric vector}
    \item{\code{Cmoth}}{a numeric vector}
  }
}

\references{
Ricci B., Franck P., Bouvier J.-C., Casado D. and Lavigne C. (2011). Effects of hedgerow characteristics on intra-orchard distribution of larval codling moth. Agriculture, Ecosystems & Environment. 140. 395-400. 10.1016/j.agee.2011.01.001. 
}



\keyword{datasets}
