\name{fitted.Bsiland}
\alias{fitted.Bsiland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compute fiited values
}
\description{
Compute fiited values for an object of class Bsiland
}

\usage{
\method{fitted}{Bsiland}(object,...)
}




%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ an object of class Bsiland}
\item{\dots}{Dots are not take into account in this version}
}

\value{
Returns the fitted values.
}

\author{
O. Martin
}
