\name{fitted.siland}
\alias{fitted.siland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compute fiited values
}
\description{
Compute fiited values for an object of class siland
}

\usage{
\method{fitted}{siland}(object,...)
}




%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ an object of class siland}
\item{\dots}{Dots are not take into account in this version}
}

\value{
Returns the fitted values.
}

\author{
O. Martin
}
