\name{landCmoth}
\alias{landCmoth}
\docType{data}
\title{
Data sets for landscape description: organic and conventional features.
}
\description{
landCmoth is a sf object and gives the positions for three landscape variables: conv, org and vine.

}
\usage{data("landCmoth")}
\format{
  A data frame with 11157 observations on the following 4 variables.
  \describe{
    \item{\code{conv}}{a numeric vector}
    \item{\code{org}}{a numeric vector}
    \item{\code{vine}}{a numeric vector}
    \item{\code{geometry}}{a sfc_MULTIPOLYGON}
  }
}

\keyword{datasets}
