\name{print.Fsiland}
\alias{print.Fsiland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
print an object of class Fsiland
}
\description{
The function print an object of class Fsiland. It gives the model used and the  estimated parameters.
}

\usage{
\method{print}{Fsiland}(x,...)
}


%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
an object of class Fsiland
}
  \item{\dots}{
the same arguments that the default print
}
}


%% ~Make other sections like Warning with \section{Warning }{....} ~


