\name{residuals.Bsiland}
\alias{residuals.Bsiland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Residuals for Bsiland
}
\description{
Compute residuals for a model estimated Bsiland function.
}

\usage{
\method{residuals}{Bsiland}(object,...)
}



%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{an object of class Bsiland}
\item{\dots}{Dots are not take into account in this version}
}
