\name{siland-package}
\alias{siland-package}
\docType{package}
\title{
\packageTitle{siland}
}

\description{
\packageDescription{siland}
}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{siland}
\packageIndices{siland}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}

\author{
\packageAuthor{siland}
Maintainer: \packageMaintainer{siland}
}




\keyword{ package }

