\name{siland-package}
\alias{siland-package}
\docType{package}
\title{
\packageTitle{siland}
}

\description{
\packageDescription{siland} This package allows for statistical 
inference about the scales at which landscape variables affect an ecological varibale. This ecological variable can be continous (mean of different measurements), an abundance (counting) or a proportion.\\
The approach is based on the method proposed by Chandler et al., Estimating the spatial sacles of landscape effects on abundance, 2016,  Landscape Ecology, 31, 1383-1394.
}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{siland}
\packageIndices{siland}
The principal function is the function siland(). Two functions for graphical displays are available: plotsiland() and plotcontri().
}

\author{
\packageAuthor{siland}
Maintainer: \packageMaintainer{siland}
}




\keyword{ package }

