\name{siland-package}
\alias{siland-package}
\docType{package}
\title{
\packageTitle{siland}
}

\description{
\packageDescription{siland} This package allows for statistical 
inference about the scales at which landscape variables affect an ecological varibale. This ecological variable can be continous (mean of different measurements), an abundance (counting) or a proportion.\\

}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{siland}
\packageIndices{siland}

}

\author{
\packageAuthor{siland}
Maintainer: \packageMaintainer{siland}
}




\keyword{ package }

